refactor: task page items
